backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
anthropic>=0.40.0
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
